перенсти слушатель кнопки ESC в открытие и закрытие модалки, перенести константы в constants.js
